feat: group.goを切り分け
