+++ router/v1/auth.go
@@ -1,7 +1,7 @@
 package v1
 
 import (
-	"painteer/handler"
+	"painteer/handler/v1"
 	"painteer/service"
 
 	"github.com/labstack/echo/v4"
@@ -10,7 +10,7 @@ import (
 func InitAuthRoutes(e *echo.Echo, authService service.AuthService) {
 	authGroup := e.Group("/auth")
 
-	authGroup.POST("/signup", handler.SignUp(authService))
-	authGroup.GET("/signin", handler.SignIn(authService))
-	authGroup.GET("/profile", handler.GetUserByID(authService))
+	authGroup.POST("/signup", v1.SignUp(authService))
+	authGroup.GET("/signin", v1.SignIn(authService))
+	authGroup.GET("/profile", v1.GetUserByID(authService))
 }
